First Pass Algorithm Implementation (#131)
* First batch of organization improvments

Sicxe_asm constructor. Skeleton of assemble. First pass split into get to start.

* Refacted variable

comp -> opcdoe

* added comment

* Fixed loading of listing_vector

for some reason read_file makes a new parser spererate than the on in the field of sicxe_asm. This is a workaround and not a proper fix.

* fixing constructor issues

Intstantiating instead of re declaring.

* Fixed constructor to use heap

* Merged in remaining code for if while assembler directive loop

* Fixed compile issues

* Fixed namespace for is_a_d

* Fixed indefinite loop

* Finished if while assembly directive

* Testing for addresses and symbol table

* Broke apart do_first_pass

* Made exception messages more assertive

* Added comments

write listing file, early error checking with cout.