Added Skeleton to Second Pass Method (#150)
* Updated git ignore

* Added method to check if string is hex

* Error checked and tested

* works

* Symtab Value is now int

This alows us to do arithmatic on values without casting back and forth. (Riggins Suggestion)

* Machine Code is now uint

Allows bitwise operators, unsinged so that it won't be interpreted as negative if format 4

* Added forward reference handler

Tested basic forward reference, and fail case.

* Improved Skeleton For Second Pass

Also Formatted Machine code for outstream operator using a ternerary in the stream statement.